+++ lib/BibSpace/Model/Tag.pm
@@ -1,31 +1,28 @@
 package Tag;
 
-use Data::Dumper;
 use utf8;
-use Text::BibTeX;    # parsing bib files
+use Text::BibTeX;
 use v5.16;
-
 use List::MoreUtils qw(any uniq first_index);
-
 use Moose;
+use MooseX::Storage;    # required for traits => DoNotSerialize
+with Storage;
 require BibSpace::Model::IEntity;
 require BibSpace::Model::ILabeled;
 with 'IEntity', 'ILabeled';
 
-use MooseX::Storage;
-with Storage('format' => 'JSON', 'io' => 'File');
+use BibSpace::Model::SerializableBase::TagSerializableBase;
+extends 'TagSerializableBase';
 
-has 'name'      => (is => 'rw', isa => 'Str');
-has 'type'      => (is => 'rw', isa => 'Int', default => 1);
-has 'permalink' => (is => 'rw', isa => 'Maybe[Str]');
+# Cast self to SerializableBase and serialize
+sub TO_JSON {
+  my $self = shift;
+  my $copy = $self->meta->clone_object($self);
+  return TagSerializableBase->meta->rebless_instance_back($copy)->TO_JSON;
+}
 
 has 'tagtype' =>
   (is => 'rw', isa => 'Maybe[TagType]', traits => ['DoNotSerialize'],);
-
-sub toString {
-  my $self = shift;
-  $self->freeze;
-}
 
 sub equals {
   my $self = shift;