+++ lib/BibSpace/Model/Type.pm
@@ -1,57 +1,20 @@
 package Type;
 
 use List::MoreUtils qw(any uniq);
-
-use Data::Dumper;
 use utf8;
-
 use v5.16;
-
 use Try::Tiny;
-
 use Moose;
-
 use Moose::Util::TypeConstraints;
 with 'IEntity';
+use BibSpace::Model::SerializableBase::TypeSerializableBase;
+extends 'TypeSerializableBase';
 
-use MooseX::Storage;
-with Storage('format' => 'JSON', 'io' => 'File');
-
-has 'our_type'    => (is => 'ro', isa => 'Str');
-has 'description' => (is => 'rw', isa => 'Maybe[Str]');
-has 'onLanding'   => (is => 'rw', isa => 'Int', default => 0);
-has 'bibtexTypes' => (
-  is      => 'rw',
-  isa     => 'ArrayRef[Str]',
-  traits  => ['Array'],
-  default => sub { [] },
-  handles => {
-    bibtexTypes_all        => 'elements',
-    bibtexTypes_add        => 'push',
-    bibtexTypes_map        => 'map',
-    bibtexTypes_filter     => 'grep',
-    bibtexTypes_find       => 'first',
-    bibtexTypes_find_index => 'first_index',
-    bibtexTypes_delete     => 'delete',
-    bibtexTypes_clear      => 'clear',
-    bibtexTypes_get        => 'get',
-    bibtexTypes_join       => 'join',
-    bibtexTypes_count      => 'count',
-    bibtexTypes_has        => 'count',
-    bibtexTypes_has_no     => 'is_empty',
-    bibtexTypes_sorted     => 'sort',
-  },
-);
-
-sub toString {
+# Cast self to SerializableBase and serialize
+sub TO_JSON {
   my $self = shift;
-  return
-      "Type: '"
-    . $self->our_type
-    . "' maps to "
-    . $self->bibtexTypes_count
-    . " bibtex types: ["
-    . join(', ', $self->bibtexTypes_all) . "]\n";
+  my $copy = $self->meta->clone_object($self);
+  return TypeSerializableBase->meta->rebless_instance_back($copy)->TO_JSON;
 }
 
 sub equals {